+++ CHANGELOG.md
@@ -9,7 +9,8 @@
  - new: `HashType` Enum of common hash types that is able to test hash values
  - new: `MagicPropertyTrait` Trait to add Magic Properties
  - new: register constants using composer `autoload`
- - new: implements `Arrayable` and `JSONable` 
+ - new: implements `Arrayable` and `JSONable`
+ - new: `Os` Operating System Enum
  - fix: minor fixes and updates
 
 ### 0.4.4 @2023 May 03